Move venstar coordinator to separate module
