+++ homeassistant/components/venstar/sensor.py
@@ -23,8 +23,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import VenstarDataUpdateCoordinator, VenstarEntity
+from . import VenstarEntity
 from .const import DOMAIN
+from .coordinator import VenstarDataUpdateCoordinator
 
 RUNTIME_HEAT1 = "heat1"
 RUNTIME_HEAT2 = "heat2"
