+++ tests/components/venstar/test_climate.py
@@ -20,7 +20,7 @@ EXPECTED_BASE_SUPPORTED_FEATURES = (
 async def test_colortouch(hass: HomeAssistant) -> None:
     """Test interfacing with a venstar colortouch with attached humidifier."""
 
-    with patch("homeassistant.components.venstar.VENSTAR_SLEEP", new=0):
+    with patch("homeassistant.components.venstar.coordinator.VENSTAR_SLEEP", new=0):
         await async_init_integration(hass)
 
     state = hass.states.get("climate.colortouch")
@@ -56,7 +56,7 @@ async def test_colortouch(hass: HomeAssistant) -> None:
 async def test_t2000(hass: HomeAssistant) -> None:
     """Test interfacing with a venstar T2000 presently turned off."""
 
-    with patch("homeassistant.components.venstar.VENSTAR_SLEEP", new=0):
+    with patch("homeassistant.components.venstar.coordinator.VENSTAR_SLEEP", new=0):
         await async_init_integration(hass)
 
     state = hass.states.get("climate.t2000")
